// apps/api/src/auth/dto/naver-user.dto.ts
export class NaverUserDto {
  resultcode: string;
  message: string;
  response: {
    age?: string;
    birthday?: string;
    birthyear?: string;
    email?: string;
    gender?: string;
    id: string;
    mobile?: string;
    name?: string;
    nickname?: string;
    profile_image?: string;
  };
}